fix/component descubrir
