+++ src/components/LinksNavbar/LinksNavbar.module.scss
@@ -16,23 +16,35 @@
   background-color: var(--white);
 }
 
-.compraClickeable, .descubrirClickeable, .ayuda {
+.compraClickeable, .descubrirClickeable{
   display: flex;
   flex-direction: column;
   padding-bottom: 0px;
-  padding-top: 120px;
+  padding-top: 130px;
   padding-left: 135px;
   flex-wrap: wrap;
   gap: 20px;
   background-color: transparent;
 }
 
-.compra, .descubrir  {
+.compra, .descubrir, .ayuda {
   display: flex;
   flex-direction: column;
   padding-left: 135px;
   gap: 20px;
 }
+
+.ayuda {
+  display: flex;
+  flex-direction: column;
+  padding-bottom: 0px;
+  padding-top: 0px;
+  padding-left: 277px;
+  flex-wrap: wrap;
+  gap: 20px;
+  background-color: transparent;
+}
+
 
 .subMenuContainer{
   display: flex;
