--- src/components/LinksNavbar/ayuda.constants.js
@@ -1,9 +1,0 @@
-const subMenuAyuda = [
-    "Comprobar el estado de un pedido",
-    "Envíos y devoluciones",
-    "Encontrar una tienda",
-    "Instrucciones de construcción",
-    "Preguntas frecuentes",
-    "Ponte en contacto con nosotros",
-    "Piezas de repuesto",
-  ];
