--- src/components/LinksNavbar/comprar.constants.js
@@ -1,101 +1,0 @@
-export const subMenusCompraClick = {
-  setsTema: [
-    "Architecture",
-    "Batman",
-    "BrickHeadz",
-    "Brick Sketches",
-    "City",
-    "Classic",
-    "Creator 3in1",
-    "Creator Expert",
-    "DC",
-    "Disney",
-    "DOTS",
-    "DUPLO",
-    "Friends",
-    "Harry Potter",
-    "Ideas",
-    "Jurassic World",
-    "LEGO Animal Crossing",
-    "LEGO DREAMZzz",
-    "LEGO Education",
-    "LEGO Icons",
-    "LEGO Indiana Jones",
-    "LEGO Super Mario",
-    "Lord of the Rings",
-    "Marvel",
-    "Minecraft",
-    "Minifiguras",
-    "Monkie Kid",
-    "NINJAGO",
-    "Powered Up",
-    "SERIOUS PLAY",
-    "Sonic the Hedgehog",
-    "Speed Champions",
-    "Spider-Man",
-    "Star Wars",
-    "Technic",
-  ],
-  Edades: ["1,5+", "4+", "6+", "9+", "13+", "18+"],
-  Precios: [
-    "Menos de 20 €",
-    "20 € - 50 €",
-    "50 € - 100 €",
-    "100 € - 200 €",
-    "Más de 200 €",
-  ],
-  Articulos: [
-    "Accesorios",
-    "Mochilas",
-    "Libros",
-    "Ropa",
-    "Llaveros",
-    "Fiambreras",
-    "Juguetes de peluche",
-    "Puzles y juegos de mesa",
-    "Juegos de rol y vestimentas",
-    "Artículos de papelería",
-    "Almacenamiento",
-    "Botellas de agua LEGO",
-  ],
-  Intereses: [
-    "Para fans adultos",
-    "Animales",
-    "Manualidades",
-    "Edifiios",
-    "Coches",
-    "Programación para niños",
-    "Dragones",
-    "Fantasía",
-    "Películas",
-    "Videojuegos",
-    "Aprender a construir",
-    "Música",
-    "Ninjas",
-    "Preescolar",
-    "Princesas",
-    "Héroes del mundo real",
-    "Robots para niños",
-    "Festividades",
-    "Espacio",
-    "Deportes",
-    "STEM",
-    "Trenes",
-    "Viajes",
-    "Series de televisión",
-    "Vehículos",
-  ],
-  PickBuild: ["Pick a Brick", "Accesorios y kits de ladrillos"],
-};
-
-export const subMenuCompra = [
-  "Exclusivos",
-  "Novedades",
-  "Los más vendidos",
-  "Decoración del hogar",
-  "Ofertas y rebajas",
-  "Tarjetas regalo",
-  "Próximamente",
-  "Encuentra la inspiración",
-  "Última oportunidad para comprar",
-];
