+++ src/components/SubMenuCompraNavbar/SubMenuCompraNavbar.jsx
@@ -5,11 +5,10 @@ import Button from "../Button/Button";
 import { FaChevronRight } from "react-icons/fa6";
 import Tipografia from "../Tipografia/Tipografia";
 
-export function SubMenuNavbar({ handleClicker }) {
+export function SubMenuCompraNavbar ({ handleClicker }) {
   const [isSubMenuOpen, setIsSubMenuOpen] = useState(false);
   const [activeSubMenu, setActiveSubMenu] = useState(null);
   const [selectedItem, setSelectedItem] = useState(false);
-  const [isDescubrirVisible, setIsDescubrirVisible] = useState(false);
 
   const subMenuRef = useRef(null);
 
@@ -106,22 +105,14 @@ export function SubMenuNavbar({ handleClicker }) {
     ],
   };
 
-  const subMenuDescubrirClick = { 
-    nuestrosValores: ["The LEGO Group", "Noticias LEGO", "Sostenibilidad"],
-    nuestrasApps: ["LEGO Builder", "App LEGO Life"],
-    nuestrasRevistas: ["Catálogos LEGO"], 
-};
-
   const handleClick = (subMenu) => {
     setIsSubMenuOpen(true);
     setActiveSubMenu(subMenu);
-    setIsDescubrirVisible(true)
   };
 
   const closeSubMenu = () => {
     setIsSubMenuOpen(false);
     setActiveSubMenu(null);
-    setIsDescubrirVisible(false)
   };
 
   useOutsideAlerter(subMenuRef, () => {
@@ -140,17 +131,6 @@ export function SubMenuNavbar({ handleClicker }) {
     "Próximamente",
     "Encuentra la inspiración",
     "Última oportunidad para comprar",
-  ];
-
-  const subMenuDescubrir = [
-    "Todos los temas LEGO",
-    "Todos los intereses LEGO",
-    "Para fans adultos",
-    "Para familias",
-    "LEGO Fortnite",
-    "LEGO Insiders",
-    "LEGO Mosaic Maker",
-    "Ideas de regalos LEGO",
   ];
 
   return (
@@ -308,46 +288,6 @@ export function SubMenuNavbar({ handleClicker }) {
         </div>
       )}
 
-{/* DESCUBRIR MENU */}
-<div ref={subMenuRef}
-        className={
-          subMenuDescubrirClick[activeSubMenu] ? styles.subMenuContainer : "none"
-        }
-      >
-<div className={styles.descubrirButtons}>
-        {/* Botones principales */}
-        <Button
-          iconRight={<FaChevronRight />}
-          texto="Nuestros valores"
-          handleClick={() => handleClick("nuestrosValores")}
-          isSubMenu
-        />
-        <Button
-          iconRight={<FaChevronRight />}
-          texto="Nuestras apps"
-          handleClick={() => handleClick("nuestrasApps")}
-          isSubMenu
-        />
-        <Button
-          iconRight={<FaChevronRight />}
-          texto="Nuestras resvistas"
-          handleClick={() => handleClick("nuestrasRevistas")}
-          isSubMenu
-        />
-        </div>
-        <div className={styles.descubrir}>
-          {subMenuDescubrir &&
-            subMenuDescubrirClick[activeSubMenu]?.map((item) => (
-              <Tipografia
-                key={item}
-                color="--black"
-                texto={item}
-                isBodyXL
-              />
-            ))}
-        </div>
-
-    </div>
     </div>
   );
 }
